Renamed actionBuilder to messageBuilder
