+++ src/messageBuilder.js
